+++ frontend/compute-system/src/components/AppStore/AppCard/interfaces/appCard.ts
